Rework viewhelpers to add resources to the current page template
All registered scripts or styles will be added to the current page template. This template is later used by the PageGenerate which reads all styles and scripts from it. This way scripts and styles defined in TypoScript will be prioritized or can be ordered.

Also using the resource data type arguments in the viewhelpers to enable some more features provided by the PageGenerator.